refactor: change the name of zustand store
