+++ apps/designer/components/Layout.tsx
@@ -4,7 +4,7 @@ import { ScriptProps } from "next/script";
 import LoadingBar from "./elements/LoadingBar";
 import Welcome from "./Welcome";
 import NewDesign from "./NewDesign";
-import useStore from "../globalStore/store";
+import useStore from "../globalStore/navigationStore";
 import SavePopup from "./SavePopup";
 const Layout: React.FC<ScriptProps> = ({ children }) => {
   // TODO: improve this type later
